+++ Projects/SysLib/Includes/SysLib/ErrorCode.hpp
@@ -3,9 +3,9 @@
 
 namespace Aiva::SysLib
 {
-    enum class ErrorCode : uint32_t
+    enum class ErrorCode : uintptr_t
     {
-        Unknown = static_cast<uint32_t>(-1),
+        Unknown = static_cast<uintptr_t>(-1),
         Success = 0,
     };
 }
